Improve the test script for LoggingHandler
